requirements.txt
streamlit==0.71.0
pycaret==2.2.2
numpy==1.19.2
pandas==1.1.2
Pillow==8.0.1
